+++ src/components/shared/icons/AppleIcon.tsx
@@ -1,0 +1,10 @@
+export const AppleLogo = () => (
+  <div className="mb-0.5 flex">
+    <svg width="16" height="21" viewBox="0 0 27 32" fill="black" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M22.2912 17.0005C22.3361 21.8432 26.5454 23.4544 26.5923 23.4748C26.5566 23.5888 25.9196 25.7709 24.3745 28.0264C23.0393 29.9755 21.6532 31.9175 19.469 31.9582C17.3236 31.9979 16.6336 30.688 14.1803 30.688C11.7281 30.688 10.9617 31.9185 8.93037 31.9979C6.82264 32.0773 5.21738 29.89 3.871 27.948C1.11912 23.9755 -0.983513 16.7216 1.83971 11.8249C3.24316 9.39337 5.75043 7.85342 8.4707 7.81474C10.5407 7.77505 12.4935 9.20507 13.7594 9.20507C15.0242 9.20507 17.398 7.48497 19.894 7.73841C20.9387 7.78217 23.872 8.15978 25.7545 10.913C25.6027 11.0056 22.2556 12.9516 22.2912 17.0005ZM18.2592 5.1094C19.3783 3.75674 20.1315 1.87379 19.9256 0C18.3122 0.0651398 16.3625 1.07379 15.2057 2.42544C14.1691 3.62238 13.261 5.53892 13.5056 7.37505C15.3035 7.51449 17.1401 6.46309 18.2592 5.1094Z"
+        fill="rgb(174 174 179)"
+      />
+    </svg>
+  </div>
+);
